requirements.txt
pdfplumber
pymupdf
pillow
pytesseract
camelot-py[cv]
tabula-py
sentence-transformers
numpy
